requirements.txt
Django==5.0.7
djangorestframework==3.14.0
django-rest-passwordreset==1.5.0
drf-spectacular==0.28.0
django-extensions==3.2.3
python-dotenv==1.0.1
PyYAML==6.0.1
celery==5.3.6
redis==6.1.0
celery==5.3.6
